pyproject.toml
[tool.poetry]
name = "LLM-Robot"
version = "0.1.0"
description = ""
authors = ["kai-yamashita <kai-yama12345679@g.ecc.u-tokyo.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
xmltodict = "^0.13.0"
pyyaml = "^6.0.1"
langchain = "^0.2.0"
openai = "^1.30.1"
langchain-community = "^0.2.0"
langchain-core = "^0.2.0"
google-api-python-client = "^2.129.0"
langchain-openai = "^0.1.7"
python-dotenv = "^1.0.1"
langchainhub = "^0.1.15"
torch = "^2.3.0"
pandas = "^2.2.2"
chromadb = "^0.5.0"
pyaudio = "^0.2.14"
torchaudio = "^2.3.0"
torchvision = "^0.18.0"
openai-whisper = {git = "https://github.com/openai/whisper.git"}


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
